+++ CHANGELOG.md
@@ -17,6 +17,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-1469 Update footer with "Accompanying documents" dropdown
 - YLP-1475 Implement the team details page for patients
 - YLP-1476 Remove care teams page
+- YLP-1494 Implement patient table for caregiver
 - YLP-1496 Add possibility for a patient to add a caregiver
 ### Engineering Use
 - YLP-1256 Split Profile component into smaller subcomponents